requirements.txt
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.2.0
Django==4.2
django-cloudinary-storage==0.3.0
executing==2.0.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
packaging==24.1
parso==0.8.4
prompt_toolkit==3.0.45
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
python-magic-bin==0.4.14
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.1
wcwidth==0.2.13
whitenoise==6.7.0
